+++ index.html
@@ -231,6 +231,6 @@
   </template>
 
 
-<script src="js/main.js"></script>
+<script type="module" src="js/main.js" ></script>
 </body>
 </html>
